64bit-arch bugfix in sha1 hash function (congratulation to the implementors of PktAnon for this masterpiece of software)
git-svn-id: https://svn.berlios.de/svnroot/repos/vermont/trunk/vermont@2178 aef3b71b-58ee-0310-9ba9-8811b9f0742f

+++ src/common/hmacsha1/sha1.h
@@ -1,17 +1,19 @@
 #ifndef __SHA1_H
 #define __SHA1_H
 
+#include <stdint.h>
+
 namespace Sha1  {
 
 #define SHA1_DIGLEN 20
 
 typedef struct {
-    unsigned long state[5];
-    unsigned long count[2];
+    uint32_t state[5];
+    uint32_t count[2];
     unsigned char buffer[64];
 } SHA1_CTX;
 
-void SHA1Transform	(unsigned long state[5], unsigned char buffer[64]);
+void SHA1Transform	(uint32_t state[5], unsigned char buffer[64]);
 void SHA1_Init		(SHA1_CTX* context);
 void SHA1_Update	(SHA1_CTX* context, unsigned char* data, unsigned int len);
 void SHA1_Final		(unsigned char digest[20], SHA1_CTX* context);
